feat: Return Github repo stars as well (#8)

+++ server/crawler/kurtosis_package_content.go
@@ -7,5 +7,5 @@ type KurtosisPackageContent struct {
 
 	PackageArguments []*StarlarkFunctionArgument
 
-	Stars int
+	Stars uint64
 }
